replace old chatflows
