+++ ThreeBotPackages/tfgrid_solutions/tfgrid_solutions/chatflows/four_to_six_gateway.py
@@ -1,37 +1,28 @@
 from Jumpscale import j
+from Jumpscale.servers.gedis.GedisChatBot import StopChatFlow
 
 
 class FourToSixGateway(j.servers.chatflow.get_class()):
-    """
-    """
-
     steps = [
+        "select_pool",
         "gateway_start",
-        "expiration_time",
         "wireguard_public_get",
         "wg_reservation",
         "wg_config",
     ]
 
-    @j.baseclasses.chatflow_step(title="Gateway")
-    def gateway_start(self):
-        self.user_form_data = {}
+    @j.baseclasses.chatflow_step(title="Pool")
+    def select_pool(self):
         user_info = self.user_info()
         j.sal.reservation_chatflow.validate_user(user_info)
+        user_info = self.user_info()
+        j.sal.reservation_chatflow.validate_user(user_info)
+        self.pool_id = j.sal.chatflow_deployer.select_pool(self)
 
-        self.gateway = j.sal.reservation_chatflow.gateway_select(self)
+    @j.baseclasses.chatflow_step(title="Gateway")
+    def gateway_start(self):
+        self.gateway = j.sal.chatflow_deployer.select_gateway(self, self.pool_id)
         self.gateway_id = self.gateway.node_id
-        self.user_form_data["Gateway"] = self.gateway_id
-
-    @j.baseclasses.chatflow_step(title="Gateway expiration time")
-    def expiration_time(self):
-        self.expiration = self.datetime_picker(
-            "Please enter solution expiration time.",
-            required=True,
-            min_time=[3600, "Date/time should be at least 1 hour from now"],
-            default=j.data.time.epoch + 3900,
-        )
-        self.user_form_data["Expiration"] = j.data.time.secondsToHRDelta(self.expiration - j.data.time.epoch)
 
     @j.baseclasses.chatflow_step(title="Wireguard public key")
     def wireguard_public_get(self):
@@ -47,64 +38,47 @@ class FourToSixGateway(j.servers.chatflow.get_class()):
         if not self.publickey:
             self.privatekey, self.publickey = j.tools.wireguard.generate_key_pair()
 
-        currencies = list()
-        farm_id = self.gateway.farm_id
-        addresses = j.sal.zosv2._explorer.farms.get(farm_id).wallet_addresses
-        for address in addresses:
-            if address.asset not in currencies:
-                currencies.append(address.asset)
-
-        if len(currencies) > 1:
-            currency = self.single_choice(
-                "Please choose a currency that will be used for the payment", currencies, default="TFT", required=True,
-            )
-        else:
-            currency = currencies[0]
-        reservation = j.sal.zosv2.reservation_create()
-        j.sal.zosv2.gateway.gateway_4to6(reservation=reservation, node_id=self.gateway_id, public_key=self.publickey)
-
-        self.resv_id = j.sal.reservation_chatflow.reservation_register_and_pay(
-            reservation, self.expiration, customer_tid=j.me.tid, currency=currency, bot=self
+        self.resv_id = j.sal.chatflow_deployer.create_ipv6_gateway(
+            self.gateway_id, self.pool_id, self.publickey, SolutionType="4to6GW"
         )
-        self.reservation_result = j.sal.reservation_chatflow.reservation_wait(self, self.resv_id)
-
-        j.sal.reservation_chatflow.reservation_save(
-            self.resv_id, f"4to6GW:{self.resv_id}", "tfgrid.solutions.4to6gw.1", self.user_form_data
-        )
-
+        success = j.sal.chatflow_deployer.wait_workload(self.resv_id, self)
+        if not success:
+            raise StopChatFlow(f"Failed to deploy workload {self.resv_id}")
+        self.reservation_result = j.sal.zosv2.workloads.get(self.resv_id).result
+        print(self.reservation_result)
         res = """
-                # Use the following template to configure your wireguard connection. This will give you access to your network.
-                ## Make sure you have <a target="_blank" href="https://www.wireguard.com/install/">wireguard</a> installed
-                Click next
-                to download your configuration
-                """
+                        # Use the following template to configure your wireguard connection. This will give you access to your network.
+                        ## Make sure you have <a target="_blank" href="https://www.wireguard.com/install/">wireguard</a> installed
+                        Click next
+                        to download your configuration
+                        """
         self.md_show(j.core.text.strip(res))
 
     @j.baseclasses.chatflow_step(title="Wireguard configuration", disable_previous=True)
     def wg_config(self):
-        cfg = self.reservation_result[0].data_json
+        cfg = self.reservation_result.data_json
         wgconfigtemplate = """\
-        [Interface]
-        Address = {{cfg.ips[0]}}
-        PrivateKey = {{privatekey}}
-        {% for peer in cfg.peers %}
-        [Peer]
-        PublicKey = {{peer.public_key}}
-        AllowedIPs = {{",".join(peer.allowed_ips)}}
-        {% if peer.endpoint -%}
-        Endpoint = {{peer.endpoint}}
-        {% endif %}
-        {% endfor %}
-        """
+            [Interface]
+            Address = {{cfg.ips[0]}}
+            PrivateKey = {{privatekey}}
+            {% for peer in cfg.peers %}
+            [Peer]
+            PublicKey = {{peer.public_key}}
+            AllowedIPs = {{",".join(peer.allowed_ips)}}
+            {% if peer.endpoint -%}
+            Endpoint = {{peer.endpoint}}
+            {% endif %}
+            {% endfor %}
+            """
         config = j.tools.jinja2.template_render(text=wgconfigtemplate, cfg=cfg, privatekey=self.privatekey.decode())
         config = j.core.text.strip(config)
 
         filename = "wg-{}.conf".format(self.resv_id)
         self.download_file(msg=f"<pre>{config}</pre>", data=config, filename=filename, html=True)
         res = f"""
-# In order to connect to the 4 to 6 gateway execute this command:
-## ```wg-quick up ./{filename}```
-                """
+    # In order to connect to the 4 to 6 gateway execute this command:
+    ## ```wg-quick up ./{filename}```
+                    """
         self.md_show(res)
