+++ ThreeBotPackages/tfgrid_solutions/tfgrid_solutions/chatflows/ubuntu_deploy.py
@@ -1,23 +1,23 @@
 from Jumpscale import j
+from Jumpscale.servers.gedis.GedisChatBot import StopChatFlow
 import math
 
 
 class UbuntuDeploy(j.servers.chatflow.get_class()):
     steps = [
         "ubuntu_start",
-        "ubuntu_network",
         "ubuntu_name",
         "ubuntu_version",
         "container_resources",
+        "select_pool",
+        "ubuntu_network",
         "container_logs",
         "public_key_get",
         "container_node_id",
-        "ubuntu_farm",
         "container_ip",
-        "expiration_time",
         "overview",
-        "container_pay",
-        "ubuntu_acess",
+        "reservation",
+        "ubuntu_access",
     ]
 
     @j.baseclasses.chatflow_step()
@@ -26,34 +26,33 @@ class UbuntuDeploy(j.servers.chatflow.get_class()):
         self.query = dict()
         self.HUB_URL = "https://hub.grid.tf/tf-bootable"
         self.IMAGES = ["ubuntu-18.04", "ubuntu-19.10", "ubuntu-20.04"]
-        self.model = j.threebot.packages.tfgrid_solutions.tfgrid_solutions.bcdb_model_get("tfgrid.solutions.ubuntu.1")
         user_info = self.user_info()
         self.user_form_data["chatflow"] = "ubuntu"
         self.md_show("# This wizard will help you deploy an ubuntu container", md=True)
         j.sal.reservation_chatflow.validate_user(user_info)
 
-    @j.baseclasses.chatflow_step(title="Network")
-    def ubuntu_network(self):
-        self.network = j.sal.reservation_chatflow.network_select(self, j.me.tid)
-
     @j.baseclasses.chatflow_step(title="Solution name")
     def ubuntu_name(self):
-        self.user_form_data["Solution name"] = j.sal.reservation_chatflow.solution_name_add(self, self.model)
+        self.solution_name = j.sal.chatflow_deployer.ask_name(self)
 
     @j.baseclasses.chatflow_step(title="Ubuntu version")
     def ubuntu_version(self):
-        self.user_form_data["Version"] = self.single_choice("Please choose ubuntu version", self.IMAGES, required=True)
+        self.version = self.single_choice("Please choose ubuntu version", self.IMAGES, required=True)
 
     @j.baseclasses.chatflow_step(title="Container resources")
     def container_resources(self):
-        form = self.new_form()
-        cpu = form.int_ask("Please add how many CPU cores are needed", default=1, required=True)
-        memory = form.int_ask("Please add the amount of memory in MB", default=1024, required=True)
-        self.rootfs_size = form.int_ask("Choose the amount of storage for your root filesystem in MiB", default=256)
-        form.ask()
-        self.user_form_data["CPU"] = cpu.value
-        self.user_form_data["Memory"] = memory.value
-        self.user_form_data["Root filesystem Size"] = self.rootfs_size.value
+        self.resources = j.sal.chatflow_deployer.ask_container_resources(self)
+
+    @j.baseclasses.chatflow_step(title="Pool")
+    def select_pool(self):
+        # FIXME: properly calculate cu and su
+        cu = self.resources["cpu"]
+        su = self.resources["disk_size"]
+        self.pool_id = j.sal.chatflow_deployer.select_pool(self, cu, su)
+
+    @j.baseclasses.chatflow_step(title="Network")
+    def ubuntu_network(self):
+        self.network_view = j.sal.chatflow_deployer.select_network(self, self.pool_id)
 
     @j.baseclasses.chatflow_step(title="Container logs")
     def container_logs(self):
@@ -63,28 +62,13 @@ class UbuntuDeploy(j.servers.chatflow.get_class()):
             default="NO",
         )
         if self.container_logs_option == "YES":
-            form = self.new_form()
-            self.channel_type = form.string_ask("Please add the channel type", default="redis", required=True)
-            self.channel_host = form.string_ask(
-                "Please add the IP address where the logs will be output to", required=True
-            )
-            self.channel_port = form.int_ask(
-                "Please add the port available where the logs will be output to", required=True
-            )
-            self.channel_name = form.string_ask(
-                "Please add the channel name to be used. The channels will be in the form NAME-stdout and NAME-stderr",
-                default=self.user_form_data["Solution name"],
-                required=True,
-            )
-            form.ask()
-            self.user_form_data["Logs Channel type"] = self.channel_type.value
-            self.user_form_data["Logs Channel host"] = self.channel_host.value
-            self.user_form_data["Logs Channel port"] = self.channel_port.value
-            self.user_form_data["Logs Channel name"] = self.channel_name.value
+            self.log_config = j.sal.chatflow_deployer.ask_container_logs(self, self.solution_name)
+        else:
+            self.log_config = {}
 
     @j.baseclasses.chatflow_step(title="Access keys")
     def public_key_get(self):
-        self.user_form_data["Public key"] = self.upload_file(
+        self.public_key = self.upload_file(
             """Please add your public ssh key, this will allow you to access the deployed container using ssh.
                     Just upload the file with the key""",
             required=True,
@@ -92,113 +76,74 @@ class UbuntuDeploy(j.servers.chatflow.get_class()):
 
     @j.baseclasses.chatflow_step(title="Container node id")
     def container_node_id(self):
-        self.var_dict = {"pub_key": self.user_form_data["Public key"]}
-        self.query["mru"] = math.ceil(self.user_form_data["Memory"] / 1024)
-        self.query["cru"] = self.user_form_data["CPU"]
-        storage_units = math.ceil(self.rootfs_size.value / 1024)
-        self.query["sru"] = storage_units
-
-        # create new reservation
-        self.reservation = j.sal.zosv2.reservation_create()
-        self.nodeid = self.string_ask(
-            "Please enter the nodeid you would like to deploy on if left empty a node will be chosen for you"
-        )
-        while self.nodeid:
-            try:
-                self.node_selected = j.sal.reservation_chatflow.validate_node(
-                    self.nodeid, self.query, self.network.currency
-                )
-                break
-            except (j.exceptions.Value, j.exceptions.NotFound) as e:
-                message = "<br> Please enter a different nodeid to deploy on or leave it empty"
-                self.nodeid = self.string_ask(str(e) + message, html=True, retry=True)
-        self.query["currency"] = self.network.currency
-
-    @j.baseclasses.chatflow_step(title="Ubuntu container farm")
-    def ubuntu_farm(self):
-        if not self.nodeid:
-            farms = j.sal.reservation_chatflow.farm_names_get(1, self, **self.query)
-            self.node_selected = j.sal.reservation_chatflow.nodes_get(1, farm_names=farms, **self.query)[0]
+        query = {
+            "cru": self.resources["cpu"],
+            "mru": math.ceil(self.resources["memory"] / 1024),
+            "mru": math.ceil(self.resources["memory"] / 1024),
+            "sru": self.resources["disk_size"],
+        }
+        self.selected_node = j.sal.chatflow_deployer.ask_container_placement(self, self.pool_id, **query)
+        if not self.selected_node:
+            self.selected_node = j.sal.chatflow_deployer.schedule_container(self.pool_id, **query)
 
     @j.baseclasses.chatflow_step(title="Container IP")
     def container_ip(self):
-        self.network_copy = self.network.copy(j.me.tid)
-        self.network_copy.add_node(self.node_selected)
-        self.ip_address = self.network_copy.ask_ip_from_node(
-            self.node_selected, "Please choose IP Address for your solution"
+        self.network_view_copy = self.network_view.copy()
+        result = j.sal.chatflow_deployer.add_network_node(
+            self.network_view.name, self.selected_node, self.network_view_copy
         )
-        self.user_form_data["IP Address"] = self.ip_address
-
-    @j.baseclasses.chatflow_step(title="Expiration time")
-    def expiration_time(self):
-        self.expiration = self.datetime_picker(
-            "Please enter solution expiration time.",
-            required=True,
-            min_time=[3600, "Date/time should be at least 1 hour from now"],
-            default=j.data.time.epoch + 3900,
-        )
-        self.user_form_data["Solution expiration"] = j.data.time.secondsToHRDelta(self.expiration - j.data.time.epoch)
+        if result:
+            for wid in result["ids"]:
+                success = j.sal.chatflow_deployer.wait_workload(wid, self)
+                if not success:
+                    raise StopChatFlow(f"Failed to add node {self.selected_node.node_id} to network {wid}")
+        free_ips = self.network_view_copy.get_node_free_ips(self.selected_node)
+        self.ip_address = self.drop_down_choice("Please choose IP Address for your solution", free_ips)
 
     @j.baseclasses.chatflow_step(title="Confirmation")
     def overview(self):
-        self.md_show_confirm(self.user_form_data)
+        self.metadata = {
+            "Solution Name": self.solution_name,
+            "Network": self.network_view.name,
+            "Node ID": self.selected_node.node_id,
+            "Pool": self.pool_id,
+            "CPU": self.resources["cpu"],
+            "Memory": self.resources["memory"],
+            "Disk Size": self.resources["disk_size"],
+            "IP Address": self.ip_address,
+        }
+        self.metadata.update(self.log_config)
+        self.md_show_confirm(self.metadata)
 
-    @j.baseclasses.chatflow_step(title="Payment", disable_previous=True)
-    def container_pay(self):
-        self.network = self.network_copy
-        self.network.update(j.me.tid, currency=self.query["currency"], bot=self)
-        container_flist = f"{self.HUB_URL}/3bot-{self.user_form_data['Version']}.flist"
-        storage_url = "zdb://hub.grid.tf:9900"
-        entry_point = "/bin/bash /start.sh"
-
-        # create container
-        cont = j.sal.zosv2.container.create(
-            reservation=self.reservation,
-            node_id=self.node_selected.node_id,
-            network_name=self.network.name,
+    @j.baseclasses.chatflow_step(title="Reservation")
+    def reservation(self):
+        container_flist = f"{self.HUB_URL}/3bot-{self.version}.flist"
+        metadata = {"SolutionType": "ubuntu", "SolutionName": self.solution_name}
+        self.resv_id = j.sal.chatflow_deployer.deploy_container(
+            pool_id=self.pool_id,
+            node_id=self.selected_node.node_id,
+            network_name=self.network_view.name,
             ip_address=self.ip_address,
             flist=container_flist,
-            storage_url=storage_url,
-            disk_type="SSD",
-            disk_size=self.rootfs_size.value,
-            env=self.var_dict,
+            cpu=self.resources["cpu"],
+            memory=self.resources["memory"],
+            disk_size=self.resources["disk_size"],
+            env={"pub_key": self.public_key},
             interactive=False,
-            entrypoint=entry_point,
-            cpu=self.user_form_data["CPU"],
-            memory=self.user_form_data["Memory"],
+            entrypoint="/bin/bash /start.sh",
+            log_config=self.log_config,
+            **metadata,
         )
-        if self.container_logs_option == "YES":
-            j.sal.zosv2.container.add_logs(
-                cont,
-                channel_type=self.user_form_data["Logs Channel type"],
-                channel_host=self.user_form_data["Logs Channel host"],
-                channel_port=self.user_form_data["Logs Channel port"],
-                channel_name=self.user_form_data["Logs Channel name"],
-            )
-        metadata = dict()
-        metadata["chatflow"] = self.user_form_data["chatflow"]
-        metadata["Solution name"] = self.user_form_data["Solution name"]
-        metadata["Version"] = self.user_form_data["Version"]
-        metadata["Solution expiration"] = self.user_form_data["Solution expiration"]
-
-        res = j.sal.reservation_chatflow.solution_model_get(
-            self.user_form_data["Solution name"], "tfgrid.solutions.ubuntu.1", metadata
-        )
-        reservation = j.sal.reservation_chatflow.reservation_metadata_add(self.reservation, res)
-        self.resv_id = j.sal.reservation_chatflow.reservation_register_and_pay(
-            reservation, self.expiration, customer_tid=j.me.tid, currency=self.query["currency"], bot=self
-        )
-
-        j.sal.reservation_chatflow.reservation_save(
-            self.resv_id, self.user_form_data["Solution name"], "tfgrid.solutions.ubuntu.1", self.user_form_data
-        )
+        success = j.sal.chatflow_deployer.wait_workload(self.resv_id, self)
+        if not success:
+            raise StopChatFlow(f"Failed to deploy workload {self.resv_id}")
 
     @j.baseclasses.chatflow_step(title="Success", disable_previous=True)
-    def ubuntu_acess(self):
+    def ubuntu_access(self):
         res = f"""\
-            # Ubuntu has been deployed successfully: your reservation id is: {self.resv_id}
-            To connect ```ssh root@{self.ip_address}``` .It may take a few minutes.
-            """
+                # Ubuntu has been deployed successfully: your reservation id is: {self.resv_id}
+                To connect ```ssh root@{self.ip_address}``` .It may take a few minutes.
+                """
         self.md_show(j.core.text.strip(res), md=True)
